[Add] Considera caso que ya se meta a un grupo en un torneo en el que ya esta

+++ WCO_API/WCO_Api/WEBModels/Tournament_Account_SWEB.cs
@@ -8,7 +8,7 @@
 
         public string acc_email { get; set; }
 
-        public int? points { get; set; } = 0;
+        public float? points { get; set; } = 0;
 
         public string GId { get; set; }
